// MadPay724-Admin/src/app/data/models/notify.ts
export interface Notify {
    enterEmail: boolean;
    enterSms: boolean;
    enterTelegram: boolean;
    exitEmail: boolean;
    exitSms: boolean;
    exitTelegram: boolean;
    ticketEmail: boolean;
    ticketSms: boolean;
    ticketTelegram: boolean;
   loginEmail: boolean;
    loginSms: boolean;
    loginTelegram: boolean;
}